Auth Context e navegação com o Router... Tem um problema na navegação, mas é porque mudei os nomes das rotas -- ajustar na develop
